# openapi.yaml
openapi: 3.0.3
info:
  title: FoodFlow 食光 API
  version: 0.1.0
  description: 家庭食材、菜单、采购和后台任务 API。数量均为最多三位小数的字符串。
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    household: {name: household, in: path, required: true, schema: {type: string, format: uuid}}
    plan: {name: plan, in: path, required: true, schema: {type: string, format: uuid}}
...
    idempotency: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 120}}
  schemas:
...
    Credentials:
      type: object
      required: [password]
      description: 推荐account统一输入手机号或邮箱；也兼容email与phone。三者只能传一个。注册需name，手机号注册还需register用途的challenge_id和code。密码限制8至72字节。
      oneOf:
        - required: [email]
          not: {anyOf: [{required: [phone]}, {required: [account]}]}
        - required: [phone]
          not: {anyOf: [{required: [email]}, {required: [account]}]}
        - required: [account]
          not: {anyOf: [{required: [email]}, {required: [phone]}]}
      properties: {account: {type: string}, email: {type: string, format: email}, phone: {type: string, pattern: '^(\+86)?1[3-9][0-9]{9}$'}, password: {type: string, minLength: 8}, name: {type: string}, code: {type: string}, challenge_id: {type: string, format: uuid}}
    SMSProof:
      type: object
      required: [phone, code, challenge_id]
      properties: {phone: {type: string}, code: {type: string, pattern: '^[0-9]{6}$'}, challenge_id: {type: string, format: uuid}, password: {type: string, description: 绑定手机号时必须提供当前密码}}
    Quantity: {type: string, pattern: '^[0-9]+(\.[0-9]{1,3})?$', example: '300.000'}
...
paths:
...
  /auth/sms/code:
    post:
      security: []
      summary: 发送注册或登录验证码
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [phone, purpose], properties: {phone: {type: string}, purpose: {type: string, enum: [register, login]}}}}}}
      responses: {'200': {description: challenge_id和expires_in及retry_after}, '429': {description: 发送限额或间隔限制}, '503': {description: 短信未配置}, '502': {description: 发送未成功确认}}
  /auth/sms/login:
    post:
      security: []
      summary: 验证短信并为已验证手机号账号签发JWT
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SMSProof'}}}}
      responses: {'200': {description: JWT会话}, '401': {description: 验证失败或手机号尚未验证}}
  /me/phone/code:
    post:
      summary: 为当前用户发送绑定验证码
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [phone], properties: {phone: {type: string}}}}}}
      responses: {'200': {description: challenge_id和expires_in及retry_after}, '429': {description: 发送限额}, '503': {description: 短信未配置}}
  /me/phone:
    post:
      summary: 当前密码和绑定验证码验证后添加手机号，拒绝换绑及其他账号占用的号码
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SMSProof'}}}}
      responses: {'204': {description: 绑定成功}, '400': {description: 验证码无效}, '401': {description: 当前密码错误}, '409': {description: 手机号不可绑定}}
  /register:
    post:
      security: []
      summary: 手机号或邮箱注册并返回7天有效JWT
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Credentials'}}}}
      responses: {'200': {description: 已注册}, '400': {description: 输入错误}}
  /login:
    post:
      security: []
      summary: 手机号或邮箱加密码登录（每IP每15分钟最多60次注册与登录）
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Credentials'}}}}
      responses: {'200': {description: 令牌}, '401': {description: 凭证错误}}
...
  /me:
    get: {summary: 获取个人资料, responses: {'200': {description: 个人资料}}}
    patch:
      summary: 修改名称、时区和提前提醒天数
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name, timezone, remind_days], properties: {name: {type: string}, timezone: {type: string}, remind_days: {type: integer, minimum: 0, maximum: 30}}}}}}
      responses: {'204': {description: 已保存}}
  /households:
    get: {summary: 列出我的家庭, responses: {'200': {description: 家庭列表}}}
    post:
      summary: 创建家庭
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name, servings], properties: {name: {type: string}, servings: {type: integer, minimum: 1, maximum: 20}, preferences: {type: string}, excluded_ingredients: {type: array, items: {type: string}}}}}}}
      responses: {'201': {description: 家庭已创建}}
  /invites/accept:
    post:
      summary: 接受指定邮箱的邀请
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [code], properties: {code: {type: string}}}}}}
      responses: {'200': {description: 已加入}}
  /recipes:
    get:
      summary: 筛选项目自编示例菜谱
      parameters: [{name: ingredient, in: query, schema: {type: string}}, {name: tag, in: query, schema: {type: string}}, {name: max_minutes, in: query, schema: {type: integer}}]
      responses: {'200': {description: 菜谱列表}}
  /ingredient-catalog:
    get:
      summary: 列出分类食材目录
      description: 需登录。返回规范名称、分类、默认单位与搜索别名，供页面单字即时提示。
      responses: {'200': {description: 食材目录}, '401': {description: 未登录}}
...
  /households/{household}:
    parameters: [{$ref: '#/components/parameters/household'}]
    get: {summary: 家庭设置, responses: {'200': {description: 家庭详情}, '404': {description: 无访问权限}}}
    patch: {summary: 修改家庭设置（仅户主）, responses: {'204': {description: 已保存}}}
  /households/{household}/members:
    parameters: [{$ref: '#/components/parameters/household'}]
    get: {summary: 家庭成员, responses: {'200': {description: 成员列表}}}
  /households/{household}/members/{user}:
    parameters: [{$ref: '#/components/parameters/household'}, {name: user, in: path, required: true, schema: {type: string, format: uuid}}]
    patch: {summary: 修改非户主角色（仅户主）, responses: {'204': {description: 已修改}}}
  /households/{household}/invites:
    parameters: [{$ref: '#/components/parameters/household'}]
    post: {summary: 创建邀请码（仅户主）, responses: {'201': {description: 邀请码}}}
  /households/{household}/inventory:
    parameters: [{$ref: '#/components/parameters/household'}]
    get: {summary: 食材汇总与批次, responses: {'200': {description: 含批次 expires_at、condition、有效余额与安全线状态}}}
  /households/{household}/today:
    parameters: [{$ref: '#/components/parameters/household'}]
    get:
      summary: 指定日期的已确认餐次
      parameters: [{name: day, in: query, required: true, schema: {type: string, format: date}}]
      responses: {'200': {description: 含 plan_id、meal_id 和逐餐完成状态的当日菜单}, '400': {description: 日期无效}}
  /households/{household}/ingredients:
    parameters: [{$ref: '#/components/parameters/household'}]
    post: {summary: 创建或更新食材, responses: {'201': {description: 食材 ID}}}
  /households/{household}/ingredients/{ingredient}:
    parameters: [{$ref: '#/components/parameters/household'}, {name: ingredient, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/idempotency'}]
    delete:
      summary: 从当前库存移除食材并审计剩余数量
      description: 在事务中锁定食材及批次，对剩余库存追加补偿流水并归零，然后将食材归档。历史批次、流水和照片保留；从目录再次入库会恢复同一食材。重复使用同一幂等键只返回原结果。
      responses: {'200': {description: 已归档及移除数量}, '403': {description: 只读成员}, '404': {description: 不属于当前家庭}, '409': {description: 幂等键内容冲突}}
  /households/{household}/catalog-stock:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/idempotency'}]
    post:
      summary: 选择目录食材并一次性创建批次入库
      description: 服务端以 catalog_id 读取规范名称、分类和单位；重放相同幂等键只返回原批次。
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [catalog_id, quantity], properties: {catalog_id: {type: string, format: uuid}, quantity: {$ref: '#/components/schemas/Quantity'}, low: {$ref: '#/components/schemas/Quantity'}, location: {type: string}, bought_on: {type: string, format: date}, expires_on: {type: string, format: date}}}}}}
      responses: {'200': {description: 食材与批次 ID}, '400': {description: 目录项或数量无效}, '409': {description: 幂等键内容冲突}}
  /households/{household}/ingredients/{ingredient}/image:
    parameters: [{$ref: '#/components/parameters/household'}, {name: ingredient, in: path, required: true, schema: {type: string, format: uuid}}]
    get: {summary: 读取家庭食材图片（需家庭成员权限）, responses: {'200': {description: JPEG、PNG 或 WebP 图片}, '404': {description: 无图片或无权限}}}
    post:
      summary: 为食材上传或更换图片（编辑者）
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [image], properties: {image: {type: string, format: binary}}}}}}
      responses: {'200': {description: 图片已保存}, '400': {description: 格式或大小无效}, '403': {description: 只读成员}}
    delete: {summary: 移除食材图片（编辑者）, responses: {'204': {description: 已移除}, '403': {description: 只读成员}}}
  /households/{household}/recipe-options:
    parameters: [{$ref: '#/components/parameters/household'}]
    post:
      summary: 按所选库存食材计算可做菜谱与依赖缺口
      description: 仅将所选食材的库存计入可用量，返回完全可做或缺 1–2 种原料的菜谱；不兼容单位返回待确认状态，不生成精确缺口。只读计算，不创建菜单或扣库存。
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [selected_ingredient_ids, servings, max_minutes], properties: {selected_ingredient_ids: {type: array, maxItems: 50, items: {type: string, format: uuid}}, servings: {type: integer, minimum: 1, maximum: 20}, max_minutes: {type: integer, minimum: 1, maximum: 240}}}}}}
      responses: {'200': {description: 按可做程度排序的菜谱与逐项需求}, '404': {description: 食材不属于该家庭}, '409': {description: 选中食材已无库存}}
  /households/{household}/clear-fridge:
    parameters: [{$ref: '#/components/parameters/household'}]
    post:
      summary: 清冰箱多菜组合建议
      description: 在授权家庭的库存与菜谱快照上，用位集匹配、到期堆与有界组合搜索给出无共享原料超分配的建议；不预留或扣减库存，提交菜单时必须重新校验。
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [selected_ingredient_ids, servings, max_minutes], properties: {selected_ingredient_ids: {type: array, minItems: 1, maxItems: 50, items: {type: string, format: uuid}}, servings: {type: integer, minimum: 1, maximum: 20}, max_minutes: {type: integer, minimum: 1, maximum: 240}}}}}}
      responses: {'200': {description: 菜谱组合、批次使用建议、原料冲突与单位疑义}, '404': {description: 食材不属于当前家庭}}
  /households/{household}/stock:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/idempotency'}]
    post:
      summary: 采购、手工录入、消耗、报损或补偿纠正
      description: batch_id 留空只适用于增加库存。扣减需要批次 ID，余额不足返回 409。
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [ingredient_id, quantity, reason], properties: {ingredient_id: {type: string, format: uuid}, batch_id: {type: string, format: uuid}, quantity: {$ref: '#/components/schemas/Quantity'}, reason: {type: string, enum: [purchase, manual, consume, waste, correction]}, note: {type: string, enum: [spoiled, expired]}, direction: {type: string, enum: [increase, decrease]}, expires_on: {type: string, format: date}, expiry_kind: {type: string, enum: [label, user, estimate, unknown]}}}}}}
      responses: {'200': {description: 库存变更}, '409': {description: 库存不足或幂等冲突}}
  /households/{household}/batches/{batch}/quantity:
    parameters: [{$ref: '#/components/parameters/household'}, {name: batch, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/idempotency'}]
    patch:
      summary: 根据当前余额纠正批次实际余量
      description: 使用预期余额作并发检查；余额变化写入不可变的补偿流水。
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [expected_quantity, target_quantity], properties: {expected_quantity: {type: string}, target_quantity: {type: string}}}}}}
      responses: {'200': {description: 纠正后余量}, '409': {description: 批次余额已变化或幂等冲突}}
  /households/{household}/batches/{batch}/cost:
    parameters: [{$ref: '#/components/parameters/household'}, {name: batch, in: path, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: 查看批次采购成本
      responses: {'200': {description: 已登记成本或 recorded=false}, '404': {description: 批次不可访问}}
    post:
      parameters: [{$ref: '#/components/parameters/idempotency'}]
      summary: 登记批次实付成本（人民币）
      description: 编辑成员确认采购成本，保存为不可变事实。后续流水按原始入库量分摊成本；此前出库不追溯定价。已有成本返回冲突，同键重试返回原结果。
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [total_cost], properties: {total_cost: {type: string, pattern: '^(0|[1-9][0-9]{0,9})(\.[0-9]{1,2})?$'}}}}}}
      responses: {'200': {description: 成本已登记}, '400': {description: 金额格式或幂等键无效}, '404': {description: 批次不存在或不属于当前家庭}, '409': {description: 成本已记录或幂等冲突}}
  /households/{household}/meals/{meal}/record:
    parameters: [{$ref: '#/components/parameters/household'}, {name: meal, in: path, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: 查看烹饪时固化的菜谱、人数与家庭时区
      responses: {'200': {description: 不可变的烹饪快照}, '404': {description: 无访问权限、未完成或历史快照缺失}}
  /households/{household}/expiring:
    parameters: [{$ref: '#/components/parameters/household'}]
    get: {summary: 七天内临期批次, responses: {'200': {description: 临期列表}}}
  /households/{household}/plans:
    parameters: [{$ref: '#/components/parameters/household'}]
    get: {summary: 最近菜单, responses: {'200': {description: 菜单列表}}}
    post:
      summary: 手工创建待确认菜单；同一餐可包含多道菜
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [meals], properties: {meals: {type: array, items: {$ref: '#/components/schemas/Meal'}}}}}}}
      responses: {'201': {description: 草案}}
  /households/{household}/plans/{plan}:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/plan'}]
    get: {summary: 菜单与实时重算食材缺口, responses: {'200': {description: 菜单详情}}}
  /households/{household}/plans/{plan}/meals/{meal}:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/plan'}, {name: meal, in: path, required: true, schema: {type: string, format: uuid}}]
    patch: {summary: 替换主菜、调整人数、变更附加菜或取消餐次, description: additional_recipe_ids 省略时保留现有附加菜，传空数组时清空；所有改动仅允许草案。接口会提高 revision，确认时必须提供最新版。, responses: {'204': {description: 已修改}}}
  /households/{household}/plans/{plan}/confirm:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/plan'}, {$ref: '#/components/parameters/idempotency'}]
    post:
      summary: 确认菜单并重算采购缺口
      responses: {'200': {description: 采购清单 ID}, '400': {description: 缺少幂等键}, '409': {description: 版本变化或幂等冲突}}
  /households/{household}/plans/{plan}/reject:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/plan'}]
    post: {summary: 拒绝草案, responses: {'204': {description: 已拒绝}}}
  /households/{household}/plans/{plan}/consume:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/plan'}, {$ref: '#/components/parameters/idempotency'}]
    post: {summary: 兼容旧客户端的整份菜单消耗；新客户端使用逐餐完成, responses: {'200': {description: 已消耗}, '409': {description: 库存不足或已有餐次完成}}}
  /households/{household}/plans/{plan}/meals/{meal}/complete:
    parameters: [{$ref: '#/components/parameters/household'}, {$ref: '#/components/parameters/plan'}, {name: meal, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/idempotency'}]
    post:
      summary: 烹饪完成并按到期时间跨批扣库
      description: 餐次日期在家庭内唯一完成。主菜和附加菜的 BOM 按人数合并，事务内锁定计划和正常、未过期批次；每批追加出库流水，低于安全线时幂等写入系统建议补货项。任一原料不足则整餐回滚。相同幂等键返回原响应。
      responses: {'200': {description: 餐次已完成}, '409': {description: 已完成、库存不足或单位不可换算}}
...
  /households/{household}/prices/benchmarks:
    parameters: [{$ref: '#/components/parameters/household'}]
    get:
      summary: 当前城市各品种最新官方监测基准
      description: 保留来源、规格、价格类型及监测起止日期；历史月度数据不等于今日现价。
      parameters:
        - {name: city, in: query, required: true, schema: {type: string, maxLength: 32}}
        - {name: category, in: query, schema: {type: string, maxLength: 32}}
      responses:
        '200':
          description: 按品种、来源、规格分别保留最新记录；未覆盖城市返回空数组
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Benchmark'}}
        '400': {description: 城市或分类无效}
        '401': {description: 未登录}
        '404': {description: 非家庭成员}
  /households/{household}/prices/dashboard:
    parameters: [{$ref: '#/components/parameters/household'}]
    get:
      summary: 食材最新官方基准
      description: 金额为十进制字符串。historical=false 仅表示北京时间当天的日监测数据，月度数据永不视为今日价。周同比要求同来源同规格监测周期整体相差 7 天，否则为 null。
      parameters:
        - {name: city, in: query, required: true, schema: {type: string, maxLength: 32}}
        - {name: ingredient_name, in: query, required: true, schema: {type: string, maxLength: 64}}
      responses:
        '200':
          description: 聚合结果，无身份字段
          content:
            application/json:
              schema:
                type: object
                properties:
                  as_of: {type: string, format: date}
                  benchmarks: {type: array, items: {$ref: '#/components/schemas/Benchmark'}}
        '400': {description: 无效城市或目录外食材}
        '401': {description: 未登录}
        '404': {description: 非家庭成员}
...
  /households/{household}/jobs/image:
    parameters: [{$ref: '#/components/parameters/household'}]
    post:
      summary: 上传 JPEG、PNG 或 WebP 并排队识别食材
      description: 最多 5 MiB。需要配置视觉模型；识别只建议名称和分类，不推断数量、单位或有效期。
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [image], properties: {image: {type: string, format: binary}}}}}}
      responses: {'202': {description: 图片识别任务 ID}, '503': {description: 未配置视觉模型}}
...